Fix: move code, remove obsolete code, fix linting errors
