+++ app/page.tsx
@@ -1,11 +1,10 @@
-import { getPresentationText } from '@/lib/contentful/rich-text';
+import { getPresentationText } from '@/lib/contentful/presentation/api';
 import ParallaxText from '@/components/parallax-text';
 export default async function Home() {
-
-  const { text } = await getPresentationText();
+  const { text, headerImage } = await getPresentationText();
   return (
     <div className="min-h-screen">
-    <ParallaxText text={text} background="/Evgeni_Leonov_foto.jpg"/>
+      <ParallaxText text={text} background={headerImage.fullUrl} />
     </div>
   );
 }
