+++ lib/contentful/presentation/types.ts
@@ -1,12 +1,12 @@
-import { CFImageFile } from "../types";
-import type { Document as CfDocument } from '@contentful/rich-text-types';
+import { ChainModifiers, Entry, EntryFieldTypes, EntrySkeletonType, LocaleCode } from 'contentful';
 
-export type PresentationFields = {
-  presentationstext?: CfDocument;
-  headerImage?: CFImageFile;
-};
+export interface TypePresentationFields {
+  presentationstext?: EntryFieldTypes.RichText;
+  headerImage?: EntryFieldTypes.AssetLink;
+}
 
-export type PresentationEntry = {
-  sys: { id: string };
-  fields: PresentationFields;
-};
+export type TypePresentationSkeleton = EntrySkeletonType<TypePresentationFields, 'presentation'>;
+export type TypePresentation<
+  Modifiers extends ChainModifiers,
+  Locales extends LocaleCode = LocaleCode,
+> = Entry<TypePresentationSkeleton, Modifiers, Locales>;
