+++ lib/contentful/rich-text.tsx
@@ -9,8 +9,6 @@ import {
 
 
 import { documentToReactComponents, type Options } from '@contentful/rich-text-react-renderer';
-import { fetchEntries } from './api';
-import { PresentationEntry } from './presentation/types';
 
 function isHyperlink(node: Block | Inline): node is Inline & { data?: { uri?: string } } {
   return node.nodeType === INLINES.HYPERLINK;
@@ -40,23 +38,3 @@ export function renderRichText(richText?: CfDocument | undefined): ReactNode {
   return documentToReactComponents(richText, options);
 }
 
-export async function getPresentationText() {
-  const response = await fetchEntries('presentation');
-  
-  return mapPresentationText(response.items);
-}
-export function mapPresentationText(entries: PresentationEntry[]): {
-  text: ReactNode;
-  headerImage: string;
-} {
-  const first = entries[0];
-  const rich = first?.fields?.presentationstext;
-  
-  const raw = first?.fields?.headerImage?.fields?.file?.url;
-  const headerImage = raw ? `http:${raw}?w=1600&fm=webp&q=80` : '';
-  
-  return {
-    text: renderRichText(rich),
-    headerImage,
-  };
-}
